chore: move Store to platform
