+++ src/platform/index.ts
@@ -5,3 +5,4 @@ export * from './shared/displayable'
 export * from './merkle-tree'
 export * from './deployment-with-auth-chain'
 export * from './entity'
+export * from './store'
