+++ src/platform/store/index.ts
@@ -1,0 +1,1 @@
+export { Store } from './store'
